backend/requirements.txt
inference-sdk
python-dotenv